chore: Update styling in App.css and UserForm.jsx
This commit updates the styling in the App.css and UserForm.jsx files. In App.css, the margin property for the button is modified to remove the auto centering. In UserForm.jsx, the background color of the container is changed to white and the color of the typography and label components is changed to black. Additionally, the color of the submit button is changed to white and the background color is changed to #00c6e6. These changes improve the visual appearance of the application.