chore: Add submitted state variable and addUser function to User component
This commit adds a new state variable `submitted` and a new function `addUser` to the User component in the Users.jsx file. The `submitted` variable is used to track whether a user has been submitted, and the `addUser` function is responsible for making a POST request to create a new user. These additions enhance the functionality of the User component.