feat: Add endpoint for creating new users
The code changes in this commit add a new endpoint `/users` to the backend server. This endpoint allows clients to send a POST request with user data to create a new user. The user data includes fields for ID, first name, last name, email, date of birth, and account type. If the user creation is successful, the server responds with a JSON object containing the newly created user. If there is an error, the server responds with a JSON object containing an error message. This addition enhances the functionality of the application by enabling the creation of new users.