feat: Add form submission functionality to AddUser page
The code changes in this commit add form submission functionality to the AddUser page. The handleSubmit function is implemented to handle the form submission event. It makes an HTTP POST request to the backend API endpoint "/users" with the form data. If the request is successful, the form data is cleared and a success message is displayed. If there is an error, appropriate error messages are displayed based on the type of error. This addition enhances the functionality of the application by allowing users to register new users through the AddUser page.