refactor: Update userSchema field names in backend/model.js
The code changes in this commit update the field names in the userSchema of the backend/model.js file. The "f_name" field is renamed to "fname" and the "l_name" field is renamed to "lname". This refactor improves the clarity and consistency of the codebase by using more descriptive field names.