refactor: Update field names in userSchema and form data variables
The code changes in this commit refactor the field names in the userSchema of the backend/model.js file and the form data variables in the frontend AddUser and UserDashBoard components. The "f_name" field is renamed to "fname" and the "l_name" field is renamed to "lname" in the userSchema. In the AddUser component, the "firstName" and "lastName" variables are renamed to "first_name" and "last_name". In the UserDashBoard component, the "f_Name" and "l_Name" variables are renamed to "f_name" and "l_name". These changes improve the clarity and consistency of the codebase by using more descriptive field names and variable names.